Add command line exit functionality to the master.
The menu gave the option 100 to exit the master. This implements that.
There is a still a problem with workers, since after exiting them via
Ctrl+C they wait until a response it heard from the master before
shutting down. If you use 100 to shut down the master you have to start
up a new instance just to shut down the workers.