Added caching capabilities to Datasets.
Added the ability to write a dataset to disk (freeing up the associated
memory used to store the content, but not the metadata), return a
Dataset from disk to memory, and check if a Dataset is currently in
memory (or conversely, on disk). When a Dataset is on disk subsequent
additions are appended to the file containing the data rather than
storing them in memory. All three functions are wrapped in the
DataReferenceManager as well so changes can be made using the Dataset
identifier instead of requiring a pointer directly to the Dataset.