Finalized HDFS integration.
Fixed the minor outlying issues (mostly with including filenames in the
Path to be opened). This code has been tested. When everything is
configured properly it works. When there are issues, such as configuring
the wrong port in the HDFS URI or the HDFS deployment has no datanodes
running, it fails silently. That isn't ideal, but it is consistent with
the rest of SEEPng.