Added CSVParser.
Added a new text input schema parser that does basic CSV parsing. One
record per line, every comma separates a new field. The stringFromBytes
function is not fully tested, but since we are focussed on the input at
the moment that isn't as important. Note that this does have to be in
the code rather than the examples because of how some of the
serialization stuff is handled.