Altered disk evictions to properly use length.
Previously the length of the buffer was put into the eviction file as an
int, but read in as a byte (I don't know why this didn't cause more
problems). It is now written and read as an int. I also tried to use the
int as read back, so we might not be limited to min_buffer_size chunks.
Or we might be. I'm not entirely sure.