Minor updates to eviction.
Update 1: Changes the freed memory from an int to a long. Since ints are
unsigned, freed data sizes between 2 and 4 GB were returning negative
freed space, which failed a >0 check that should have passed.
Update 2: Allows eviction even if the first buffer is at a position
greater than 0. I'm not sure why that was added, but it was causing
problems when the input (synthetic) dataset size was too large to fit
into memory.